requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
python-multipart
google-generativeai
slowapi
pypdf